Se implementa el template de registro en el servicio y se refactoriza.
